requirements.txt
openai>=1.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
PyYAML>=6.0
edge-tts>=6.0.0
aiohttp>=3.8.0
bilibili-api-python
moviepy